Change DecimalPlaces from int to byte, and correct some minor bugs

+++ PersonalFinanceApp.Application/Features/Currencies/Commands/CreateCurrency/CreateCurrencyCommand.cs
@@ -10,6 +10,6 @@ public class CreateCurrencyCommand : IRequest<byte>
 {
     public string Code { get; set; } = string.Empty;
     public string Name { get; set; } = string.Empty;
-    public int DecimalPlaces { get; set; }
+    public byte DecimalPlaces { get; set; }
     public string Symbol { get; set; } = string.Empty;
 }
